Updated frontend according to ESLint
