--- frontend/src/Calendar.js
@@ -1,9 +1,0 @@
-import styles from './Calendar.module.css';
-
-function Calendar() {
-    return(
-        <text>Hello Calendar</text>
-    );
-}
-
-export default Calendar;
