+++ frontend/src/Calendar.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+// import styles from './Calendar.module.css';
+
+function Calendar() {
+  return (
+    <text>Hello Calendar</text>
+  );
+}
+
+export default Calendar;
