--- frontend/src/List.js
@@ -1,9 +1,0 @@
-import styles from './List.module.css';
-
-function List() {
-    return(
-        <text>Hello List</text>
-    );
-}
-
-export default List;
