+++ frontend/src/List.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+// import styles from './List.module.css';
+
+function List() {
+  return (
+    <text>Hello List</text>
+  );
+}
+
+export default List;
