--- frontend/src/Today.js
@@ -1,9 +1,0 @@
-import styles from './Today.module.css'
-
-function Today() {
-  return(
-      <text>Hello Today</text>
-  );
-}
-
-export default Today;
