+++ frontend/src/Today.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+// import styles from './Today.module.css';
+
+function Today() {
+  return (
+    <text>Hello Today</text>
+  );
+}
+
+export default Today;
